requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
allure-pytest==2.13.5
apscheduler==3.11.0
asyncssh==2.21.0
celery==5.3.1
colorlog==6.8.2
configobj==5.0.9
ddddocr==1.4.11
docx2pdf==0.1.8
eventlet==0.39.0
faker==13.4.0
fpdf==1.7.2
hypercorn==0.17.3
jsonpath==0.82
locust==2.31.5
loguru==0.7.2
nest-asyncio==1.6.0
openpyxl==3.1.3
pandas==2.1.4
pip-chill==1.0.3
playwright==1.28.0
py-ulid==1.0.3
pycodestyle==2.13.0
pycryptodome==3.21.0
pymysql==1.1.1
pypdf2==3.0.1
pytest-asyncio==0.25.3
pytest-order==1.3.0
python-docx==1.1.2
python-pptx==1.0.2
pyyaml==6.0.1
redis==4.5.2
ruamel.yaml==0.18.10
sshtunnel==0.4.0
ulid==1.1
websockets==12.0
python-jenkins==1.8.2